refactor: extract the functions from layout to mixins
