+++ lib/default-theme/mixins/index.js
@@ -1,0 +1,9 @@
+import nprogressMixin from './nprogress'
+import scrollListenerMixin from './scrollListener'
+import updateHeadMixin from './updateHead'
+
+export {
+  nprogressMixin,
+  scrollListenerMixin,
+  updateHeadMixin
+}
