+++ lib/default-theme/mixins/nprogress.js
@@ -1,0 +1,20 @@
+import Vue from 'vue'
+import nprogress from 'nprogress'
+import { pathToComponentName } from '@app/util'
+
+export default {
+  mounted () {
+    nprogress.configure({ showSpinner: false })
+
+    this.$router.beforeEach((to, from, next) => {
+      if (to.path !== from.path && !Vue.component(pathToComponentName(to.path))) {
+        nprogress.start()
+      }
+      next()
+    })
+
+    this.$router.afterEach(() => {
+      nprogress.done()
+    })
+  }
+}
